Move StreamController for day selection up to HomePage from WeekViewPager
