+++ lib/pages/home.dart
@@ -1,4 +1,7 @@
+import 'dart:async';
+
 import 'package:flutter/material.dart';
+import 'package:habittracker/util/day.dart';
 import 'package:habittracker/widgets/database_tester.dart';
 import 'package:habittracker/widgets/fab_menu.dart';
 import 'package:habittracker/widgets/week_view_pager.dart';
@@ -11,6 +14,22 @@ class HomePage extends StatefulWidget {
 }
 
 class _HomePageState extends State<HomePage> {
+  StreamController<DateTime> _selectionController;
+
+  @override
+  void initState() {
+    super.initState();
+    _selectionController = StreamController<DateTime>.broadcast();
+    _selectionController
+        .add(DateTime.parse(Day.formatter.format(DateTime.now())));
+  }
+
+  @override
+  void dispose() {
+    super.dispose();
+    _selectionController.close();
+  }
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -22,7 +41,7 @@ class _HomePageState extends State<HomePage> {
           children: <Widget>[
             SizedBox(
                 height: MediaQuery.of(context).size.height / 5,
-                child: WeekViewPager()),
+                child: WeekViewPager(selectionController: _selectionController,)),
             DatabaseTester(),
           ],
         ),
