+++ lib/widgets/week_view_pager.dart
@@ -1,27 +1,18 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
-import 'package:habittracker/util/day.dart';
 
 import 'week_view.dart';
 
 class WeekViewPager extends StatefulWidget {
-  WeekViewPager();
+  final StreamController<DateTime> selectionController;
+  WeekViewPager({this.selectionController});
 
   @override
   _WeekViewPagerState createState() => new _WeekViewPagerState();
 }
 
 class _WeekViewPagerState extends State<WeekViewPager> {
-  StreamController<DateTime> _selectionController;
-
-  @override
-  void initState() {
-    super.initState();
-    _selectionController = StreamController<DateTime>.broadcast();
-    _selectionController
-        .add(DateTime.parse(Day.formatter.format(DateTime.now())));
-  }
 
   Widget build(BuildContext context) {
     return PageView.builder(
@@ -29,15 +20,9 @@ class _WeekViewPagerState extends State<WeekViewPager> {
       itemBuilder: (BuildContext context, int index) {
         return WeekView(
           lastDay: DateTime.now().add(Duration(days: (index - 2048) * 7)),
-          selectionController: _selectionController,
+          selectionController: widget.selectionController,
         );
       },
     );
-  }
-
-  @override
-  void dispose() {
-    super.dispose();
-    _selectionController.close();
   }
 }
